requirements/pypi.txt
# Django
django
pytz

# tools
django-click
enum34
nkf
PyYAML

# REST
djangorestframework
rest-pandas
furl
requests
django-fullurl

# Backend
django-storages

# Text
markdown
py-gfm

# HTML
beautifulsoup4

# Excel
openpyxl
xlwt

# AWS
awscli
